+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Fixed spectators sometimes spawning with health.
+
 ## [v1.21.11-0.1.1] - 2026-01-17
 
 ### Added