+++ CHANGELOG.md
@@ -7,9 +7,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [v1.21.11-0.1.2] - 2026-02-07
+
+### Added
+
+- Added Fabric support
+
 ### Fixed
 
-- Fixed spectators sometimes spawning with health.
+- Fixed spectators sometimes spawning with health
 - Fixed player-hurt sound inconsistency
 
 ## [v1.21.11-0.1.1] - 2026-01-17